perma_web/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
-e git+git://github.com/rebeccacremona/createsend-python.git@7cd8be21f89fa7bb61dafb31da9ad4a64058bf5b#egg=createsend-python
-e git+git://github.com/internetarchive/warcprox.git@f79e74482339f8c06880b28e889be694b30d4575#egg=warcprox
amqp==1.4.9               # via kombu
anyjson==0.3.3            # via kombu
apipkg==1.4               # via execnet
argh==0.26.2              # via watchdog
args==0.1.0               # via clint
asn1crypto==0.22.0        # via cryptography
attrs==17.2.0             # via hypothesis
bcrypt==3.1.4             # via paramiko
beautifulsoup4==4.5.1
billiard==3.3.0.23        # via celery
boto3==1.4.4
boto==2.48.0
botocore==1.5.49          # via boto3, s3transfer
brotlipy==0.6.0           # via pywb
celery==3.1.25
certauth==1.1.3
certifi==2017.7.27.1      # via requests
cffi==1.10.0              # via bcrypt, brotlipy, cryptography, pynacl
chardet==3.0.4            # via pywb, requests
click==6.7                # via pip-tools
clint==0.5.1              # via internetarchive
coverage==4.3.4
cryptography==2.1
cssselect==1.0.1          # via pyquery
django-admin-smoke-tests==0.3.0
django-crispy-forms==1.6.1
django-filter==1.0.2
django-model-utils==2.6.1
django-mptt==0.8.7
django-ratelimit==1.1.0
django-redis==4.4.3
django-settings-context-processor==0.2
django-simple-history==1.9.0
django-sslserver==0.14
django-storages==1.5.2
django-taggit==0.22.2
django-webpack-loader==0.5.0
django==1.11.13
djangorestframework==3.6.2
docopt==0.6.2             # via internetarchive
docutils==0.13.1          # via botocore
easyprocess==0.2.3        # via pyvirtualdisplay
execnet==1.4.1            # via pytest-xdist
fabric3==1.13.1.post1
fakeredis==0.7.0
first==2.0.1              # via pip-tools
flake8==2.5.4
gevent==1.0.2
greenlet==0.4.12          # via gevent
gunicorn==19.3.0
hypothesis==3.31.2
idna==2.5                 # via cryptography, requests, tldextract
internetarchive==1.0.10
jinja2==2.8.1             # via pywb
jmespath==0.9.2           # via boto3, botocore
jsmin==2.0.9
jsonpatch==0.4            # via internetarchive
kombu==3.0.37             # via celery
linkheader==0.4.3
lxml==3.7.3               # via pyquery
markupsafe==1.0           # via jinja2
mccabe==0.4.0             # via flake8
mock==2.0.0
mysqlclient==1.3.12
netaddr==0.7.12
newrelic==2.66.0.49
paramiko==2.3.1           # via fabric3
pathtools==0.1.2          # via watchdog
pbr==2.0.0                # via mock
pep8==1.7.0               # via flake8
pexpect==3.1
pillow==3.3.2
pip-tools==1.11.0
py==1.4.33                # via pytest, pytest-xdist
pyasn1==0.3.7             # via paramiko
pycparser==2.17           # via cffi
pyflakes==1.0.0           # via flake8
pynacl==1.1.2
pyopenssl==16.2.0         # via certauth, requests
pyquery==1.2.17
pyscss==1.3.4
pytest-cov==2.5.1
pytest-django==3.1.2
pytest-xdist==1.15.0
pytest==3.0.7
python-dateutil==2.2
pytz==2018.3
pyvirtualdisplay==0.1.5
pywb==0.33.2
pyyaml==3.10              # via pywb, watchdog
redis==2.10.5
requests-file==1.4.1      # via tldextract
requests[security]==2.18.4
s3transfer==0.1.10        # via boto3
sauceclient==1.0.0
schema==0.5.0             # via internetarchive
selenium==2.47.3
six==1.10.0               # via bcrypt, cryptography, django-admin-smoke-tests, fabric3, internetarchive, mock, pip-tools, pynacl, pyopenssl, pyscss, python-dateutil, pywb, requests-file, surt, warcio
smhasher==0.150.1
sorl-thumbnail==12.3
sqlparse==0.2.4
surt==0.3.0               # via pywb
tempdir==0.6
tldextract==2.0.2         # via surt
tqdm==4.11.2
ua-parser==0.7.1
urllib3==1.22             # via requests
wand==0.4.4
warcio==1.3.3
warctools==4.10.0
watchdog==0.8.3           # via pywb
webencodings==0.5.1       # via pywb
werkzeug==0.11.1
whitenoise==3.2.2
